add schedyield setaffinity
