+++ src/sst/elements/vanadis/os/callev/voscallall.h
@@ -17,7 +17,7 @@
 #define _H_VANADIS_CALL_ALL
 
 #include "os/callev/voscallaccessev.h"
-#include "os/callev/voscallschedsetaffinity.h"
+#include "os/callev/voscallsetaffinity.h"
 #include "os/callev/voscallgetaffinity.h"
 #include "os/callev/voscallschedyield.h"
 #include "os/callev/voscallcheckpoint.h"
